feat: rename generate-pr-description command to generate-pull-request-description and improve pull request description generation command
- Added a new command to generate pull request descriptions
- Improved the pull request description generation process by adding a schema for the title and description
- Added input validation to ensure that the repository is in a clean state before generating a pull request description
- Added error handling to catch and report any errors that occur during the pull request description generation process
